@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe 'Checkout', js: true do
@@ -22,7 +24,7 @@ describe 'Checkout', js: true do
   context 'as admin user' do
     stub_authorization!
 
-    it 'should calculate and display handling for one item on payment step and allow full checkout' do
+    it 'calculates and display handling for one item on payment step and allow full checkout' do
       add_to_cart('RoR Mug')
       click_button 'Checkout'
 
@@ -31,11 +33,11 @@ describe 'Checkout', js: true do
       fill_in_address(alabama_address)
       click_button 'Save and Continue'
       click_button 'Save and Continue'
-      # page.should have_content("Handling $1.90") # TODO: Diagnose missing labels in capybara
-      expect(page).to have_content(/Order Total:\s\$21.90/)
+      click_button 'Save and Continue'
+      expect(page).to have_content("Handling: $1.90")
+      expect(page).to have_content('$21.90')
 
-      click_on 'Save and Continue'
-      click_on 'Place Order'
+      click_button 'Place Order'
       expect(current_path).to match(spree.order_path(Spree::Order.last))
 
       # Verify handling fee from backend
@@ -44,13 +46,12 @@ describe 'Checkout', js: true do
       expect(page.find('fieldset#order-total')).to have_content(/Order Total\s\$21.90/)
       expect(page.find('.js-order-shipment-adjustments')).to have_content(/Handling:\s\$1.90/)
 
-      # TODO: For some reason this override doesn't display in capybara
-      # expect(page.find('dl.additional-info')).to have_content("Total:$21.90")
-      # expect(page.find('dl.additional-info')).to have_content("Handling: $1.90")
+      expect(page.find('dl.additional-info')).to have_content("Total:\n$21.90")
+      expect(page.find('dl.additional-info')).to have_content("Handling:\n$1.90")
     end
   end
 
-  it 'should calculate and display handling for multiple items on payment step and allow full checkout' do
+  it 'calculates and display handling for multiple items on payment step and allow full checkout' do
     add_to_cart('RoR Mug')
     add_to_cart('RoR Mug')
     add_to_cart('RoR Mug')
@@ -61,7 +62,7 @@ describe 'Checkout', js: true do
     fill_in_address(alabama_address)
     click_button 'Save and Continue'
     click_button 'Save and Continue'
-    # page.should have_content("Handling $2.70") # TODO: Diagnose missing labels in capybara
+    expect(page).to have_content("Handling: $2.70")
     expect(page).to have_content(/Order Total:\s\$42.70/)
 
     click_on 'Save and Continue'
@@ -69,7 +70,7 @@ describe 'Checkout', js: true do
     expect(current_path).to match(spree.order_path(Spree::Order.last))
   end
 
-  it 'should update the handling fee when cart contents change' do
+  it 'updates the handling fee when cart contents change' do
     add_to_cart('RoR Mug')
     click_button 'Checkout'
 
@@ -78,14 +79,14 @@ describe 'Checkout', js: true do
     fill_in_address(alabama_address)
     click_button 'Save and Continue'
     click_button 'Save and Continue'
-    # page.should have_content("Handling $1.90") # TODO: Diagnose missing labels in capybara
+    expect(page).to have_content("Handling: $1.90")
     expect(page).to have_content(/Order Total:\s\$21.90/)
 
     add_to_cart('RoR Mug')
     click_button 'Checkout'
     click_button 'Save and Continue'
     click_button 'Save and Continue'
-    # page.should have_content("Handling $2.30") # TODO: Diagnose missing labels in capybara
+    expect(page).to have_content("Handling: $2.30")
     expect(page).to have_content(/Order Total:\s\$32.30/)
   end
 
