@@ -1,5 +1,7 @@
+# frozen_string_literal: true
+
 Spree::Calculator::Shipping::FlatRate.class_eval do
-  def compute_shipment(shipment)
+  def compute_shipment(_shipment)
     preferred_amount
   end
 end
